+++ Sources/PSParseHTML/Models/TableMetadata.cs
@@ -13,11 +13,11 @@ namespace PSParseHTML;
 /// </summary>
 public class TableMetadata {
     public int TableIndex { get; set; }
+    public bool IsVisible { get; set; } = true;
     public string? Id { get; set; }
     public string? Classes { get; set; }
     public Dictionary<string, string> Attributes { get; set; } = new();
     public int RowCount { get; set; }
     public int ColumnCount { get; set; }
     public List<string> Headers { get; set; } = new();
-    public bool IsVisible { get; set; } = true;
 }
