+++ Sources/PSParseHTML/Models/TableParseResult.cs
